+++ geotrek/settings/base.py
@@ -621,5 +621,7 @@ PASSWORD_HASHERS = [
     'django.contrib.auth.hashers.UnsaltedMD5PasswordHasher',  # Used for extern authent
 ]
 
-FORMAT_LINE_CODE = "{signagecode}-{bladenumber}-{linenumber}"
+BLADE_CODE_TYPE = int
+BLADE_CODE_FORMAT = "{signagecode}-{bladenumber}"
+LINE_CODE_FORMAT = "{signagecode}-{bladenumber}-{linenumber}"
 SHOW_EXTREMITIES = False