+++ docs/changelog.rst
@@ -9,6 +9,8 @@ CHANGELOG
 
 - When updating interventions, stake field is no more required
 - Fix duplicates in year filters in intervention module
+- Configurable blade code
+- Allow letters in blade number
 
 
 2.24.5 (2019-03-06)